modules/bspwm-workspace-preview/requirements.txt
tk
pillow
pynput
configparser
PyAutoGUI